Run only on specific role and adapt to Red Hat osp10
The heat template was not working on osp10 cause it was using old resources. Also the puppet code was tried to be run on role as ceph, swift what is not expected and was creating failure at deployment.

Tested and works on newton (Red Hat osp10) and ACI 2.2(2e)

Signed-off-by: Cyril Lopez <cylopez@redhat.com>

fix noironetworks/aci-tripleo-patch#1

Signed-off-by: Cyril Lopez <cylopez@redhat.com>